requirements.txt
torch==2.2.1torchvision==0.17.1numpy==1.26.4pillow==10.2.0pyyaml==6.0.1tqdm==4.66.2scikit-learn==1.3.2matplotlib==3.8.3streamlit==1.31.1opencv-python==4.9.0.80